An icon handle is a Long, not a GL texture name
IconRegion.texture began as an OpenGL texture name, which is a 32-bit integer,
because the host that draws these happens to be Minecraft and that is what
Minecraft used to hand out.

From Minecraft 1.21.6 it does not. The game's Blaze3D layer owns texture handles
behind an abstraction, they are 64-bit, and on a Vulkan backend there is no GL
name to be had at all. A host that could only answer with an Int would be pinned
to an old renderer for a reason that has nothing to do with this editor -- which
is the sort of thing a boundary type exists to prevent, not to cause.

Widening costs nothing on either side. ImGui's draw list has always taken a Long
here, so the two conversions that sat at the addImage call sites are simply gone;
a host that still has an Int widens it once where it produces the handle.

Found while porting a downstream host to a second mod loader: it was the only
place in this project that named a graphics API at all.

+++ editor-graph/src/main/kotlin/dev/ziggle/vscript/editor/graph/ValuePicker.kt
@@ -233,7 +233,7 @@ object ValuePicker {
                         // Fitted INSIDE the row rather than scaled to fill it: a wiki render is a tall
                         // portrait and stretching one to a square makes every NPC look melted.
                         val box = ROW_H - 4f
-                        dl.addImage(region.texture.toLong(), lx + 4f, ry + 2f, lx + 4f + box, ry + 2f + box, region.u0, region.v0, region.u1, region.v1)
+                        dl.addImage(region.texture, lx + 4f, ry + 2f, lx + 4f + box, ry + 2f + box, region.u0, region.v0, region.u1, region.v1)
                     }
                     textX = lx + ICON_COL
                 }
@@ -375,7 +375,7 @@ object ValuePicker {
         val y1 = y0 + PREVIEW
         dl.addRectFilled(x0 - 4f, y0 - 4f, x1 + 4f, y1 + 4f, PANEL_BG, 6f, ImDrawFlags.RoundCornersAll)
         dl.addRect(x0 - 4f, y0 - 4f, x1 + 4f, y1 + 4f, BORDER, 6f, ImDrawFlags.RoundCornersAll, 1f)
-        dl.addImage(region.texture.toLong(), x0, y0, x1, y1, region.u0, region.v0, region.u1, region.v1)
+        dl.addImage(region.texture, x0, y0, x1, y1, region.u0, region.v0, region.u1, region.v1)
     }
 
     private const val PREVIEW = 96f
